Added Piano and Violin implementation

+++ src/main/java/org/example/Strings.java
@@ -1,5 +1,5 @@
 package org.example;
 
 public interface Strings {
-    byte strings();
+    short strings();
 }
